Fixed a potential overflow, thanks to Martin Alderson ! (DIRSERVER-933)
git-svn-id: https://svn.apache.org/repos/asf/directory/apacheds/trunk@539036 13f79535-47bb-0310-9956-ffa450edef68

+++ core/src/main/java/org/apache/directory/server/core/partition/impl/btree/ExpressionEnumerator.java
@@ -247,7 +247,7 @@ public class ExpressionEnumerator implements Enumerator
         for ( int ii = 0; ii < children.size(); ii++ )
         {
             ExprNode child = children.get( ii );
-            value = ( ( Long ) child.get( "count" ) ).intValue();
+            value = ( Long ) child.get( "count" );
             minValue = Math.min( minValue, value );
 
             if ( minValue == value )
